ncs_pvt-survey-backend-ba8515c0019a/requirements.txt
boto3==1.38.15
openai==1.78.1
fastapi==0.115.12
mailersend==2.0.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
uvicorn==0.34.2
APScheduler==3.11.0
wordcloud==1.9.4